Extract shared address request limit logic
